Update Layout and add Header Component
