+++ src/components/Layout/Layout.component.jsx
@@ -1,47 +1,18 @@
 import React, { useState } from 'react';
-import { IoLogoYoutube } from 'react-icons/io5';
+import Header from '../Header';
 import Menu from '../Menu';
-import Searchbar from '../Searchbar';
-import UserImage from '../UserImage';
-import {
-  Container,
-  Header,
-  HeaderLeft,
-  HeaderRight,
-  HeaderCenter,
-  MainContent,
-  MenuHamburger,
-  Logo,
-} from './Layout.styles';
+import { Container, MainContent } from './Layout.styles';
 
 const Layout = ({ children }) => {
   const [isMenuExpanded, setIsMenuExpanded] = useState(false);
 
-  const toggleMenu = () => {
+  const handleToggleMenu = () => {
     setIsMenuExpanded(!isMenuExpanded);
   };
 
   return (
     <Container isMenuExpanded={isMenuExpanded}>
-      <Header data-testid="header">
-        <HeaderLeft>
-          <MenuHamburger
-            isMenuExpanded={isMenuExpanded}
-            onClick={toggleMenu}
-            data-testid="button-menu"
-          />
-          <Logo to="/">
-            <IoLogoYoutube />
-            <span>YuTuv</span>
-          </Logo>
-        </HeaderLeft>
-        <HeaderCenter>
-          <Searchbar />
-        </HeaderCenter>
-        <HeaderRight>
-          <UserImage />
-        </HeaderRight>
-      </Header>
+      <Header isMenuExpanded={isMenuExpanded} onToggleMenu={handleToggleMenu} />
       <Menu isExpanded={isMenuExpanded} />
       <MainContent data-testid="main-content">{children}</MainContent>
     </Container>
