+++ src/components/Layout/Layout.test.jsx
@@ -1,4 +1,4 @@
-import { fireEvent, queryByTestId, render, screen } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
 import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import renderer from 'react-test-renderer';
@@ -7,7 +7,7 @@ import SearchProvider from '../../providers/Search';
 import Layout from './Layout.component';
 
 describe('Layout component', () => {
-  it('Should check all Menu Items in Layout', () => {
+  it('Should check main components in Layout', () => {
     render(
       <BrowserRouter>
         <AuthProvider>
@@ -17,27 +17,10 @@ describe('Layout component', () => {
         </AuthProvider>
       </BrowserRouter>
     );
-    const totalMenuItems = screen.getAllByRole('listitem').length;
 
-    expect(totalMenuItems).toBe(4);
-    expect(screen.getByText('Home')).toBeInTheDocument();
-  });
-
-  it('Should expand menu size on menu button click', () => {
-    const { container } = render(
-      <BrowserRouter>
-        <AuthProvider>
-          <SearchProvider>
-            <Layout />
-          </SearchProvider>
-        </AuthProvider>
-      </BrowserRouter>
-    );
-
-    const button = queryByTestId(container, 'button-menu');
-    const initialClassName = button.className;
-    fireEvent.click(button);
-    expect(initialClassName).not.toBe(button.className);
+    expect(screen.queryByTestId('header')).toBeInTheDocument();
+    expect(screen.queryByTestId('menu')).toBeInTheDocument();
+    expect(screen.queryByTestId('main-content')).toBeInTheDocument();
   });
 
   it('Validates Layout snapshot', () => {
